Move the reva admin CLI into its own cmd/reva/admin package
